requirements.txt
pyyaml
sqlalchemy
psycopg2
pandas
numpy
plotly
missingno
seaborn